server/requirements.txt
requests==2.18.4
peewee==3.1.2
networkx==2.1
grpcio==1.10.0
colorlog==3.1.2
